+++ app/src/main/java/ir/saa/android/mt/model/daos/RegionDao.java
@@ -38,7 +38,7 @@ public interface RegionDao {
     void insertRegion(Region region);
 
     @Insert(onConflict =IGNORE)
-    List<Integer> insertRegions(List<Region> regions);
+    List<Long> insertRegions(List<Region> regions);
 
     @Update
     void updateRegion(Region region);
